Add frontend for login using ejs
Summary:
Finished yashsriv's backend with an ejs based front-end (login page and a test profile page for a successful login).

How to Test:
In the Server directory:
```
node app
```

Then direct your browser to localhost:8000 and test it out.
If password is correct, you're redirected to a test-profile page listing your access rights and username.

If your input is incorrect, you will be directed again to the /login page, with a toast displaying what went wrong on the top right.

Reason for Using EJS:
I wanted a decent way to manage views as the content would change with the user.
This seemed to be a quick and simple way to do that. Also, it looks quite neat.

Pool:
```
 __  __                 _   _
|  \/  | __ _ _ __ __ _| |_| |__   __ _ ___
| |\/| |/ _` | '__/ _` | __| '_ \ / _` / __|
| |  | | (_| | | | (_| | |_| | | | (_| \__ \
|_|  |_|\__,_|_|  \__,_|\__|_| |_|\__,_|___/

```